+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/elements/CommandMenuSubMenu.java
@@ -44,6 +44,7 @@ public class CommandMenuSubMenu {
     private boolean useOrgColour;
     private boolean useBossColour;
     private boolean useHostileColour;
+    private boolean useFixedHeader;
     private final boolean autoResize;
 
     public static final Color ORG_COLOUR = new Color(204, 204, 204);
@@ -66,7 +67,7 @@ public class CommandMenuSubMenu {
         private int width = 70;
         private int height = 15;
         private final List<CommandMenuItem.Builder> children;
-        private boolean useOrgColour, useBossColour, useHostileColour, autoResize;
+        private boolean useOrgColour, useBossColour, useHostileColour, autoResize, useFixedHeader;
         private CommandMenuSubMenu parent;
         public Builder(ResourceLocation id, Component title) {
             this.id = id;
@@ -130,6 +131,11 @@ public class CommandMenuSubMenu {
             return this;
         }
 
+        public Builder fixedHeader() {
+            this.useFixedHeader = true;
+            return this;
+        }
+
         public CommandMenuSubMenu build() {
             return buildWithParent(null);
         }
@@ -174,6 +180,7 @@ public class CommandMenuSubMenu {
         this.useHostileColour = builder.useHostileColour;
         this.useBossColour = builder.useBossColour;
         this.autoResize = builder.autoResize;
+        this.useFixedHeader = builder.useFixedHeader;
     }
 
     public int getMaxY() {
@@ -485,9 +492,13 @@ public class CommandMenuSubMenu {
             RenderSystem.enableBlend();
             guiGraphics.pose().translate(0, 0, getZ());
             guiGraphics.setColor(getColour().getRed() / 255F, getColour().getGreen() / 255F, getColour().getBlue() / 255F, 1);
-            guiGraphics.blit(getTexture(), getX(), getY(), 0, 0, 4, getHeight());
-            guiGraphics.blit(getTexture(), getX()+4, getY(), getWidth()-14, getHeight(), 4, 0, 1, getHeight(),256, 256);
-            guiGraphics.blit(getTexture(), getX()+getWidth()-10, getY(), 60, 0, 10, getHeight());
+            if (useFixedHeader) {
+                guiGraphics.blit(getTexture(), getX(), getY(), 0, 70, 70, 15);
+            } else {
+                guiGraphics.blit(getTexture(), getX(), getY(), 0, 0, ModConfigs.cmHeaderEndLWidth, getHeight());
+                guiGraphics.blit(getTexture(), getX() + ModConfigs.cmHeaderEndLWidth, getY(), getWidth() - (ModConfigs.cmHeaderEndLWidth + ModConfigs.cmHeaderEndRWidth), getHeight(), ModConfigs.cmHeaderEndLWidth + 1, 0, 1, getHeight(), 256, 256);
+                guiGraphics.blit(getTexture(), getX() + getWidth() - ModConfigs.cmHeaderEndRWidth, getY(), ModConfigs.cmHeaderEndLWidth + 3, 0, ModConfigs.cmHeaderEndRWidth, getHeight());
+            }
             if(ModConfigs.cmHeaderTextVisible) {
                 Color textColour = isActive() ? titleColour : titleColour.darker().darker();
                 guiGraphics.setColor(textColour.getRed() / 255F, textColour.getGreen() / 255F, textColour.getBlue() / 255F, 1);