+++ src/main/java/online/kingdomkeys/kingdomkeys/config/ClientConfig.java
@@ -12,10 +12,8 @@ import online.kingdomkeys.kingdomkeys.KingdomKeys;
  */
 public class ClientConfig {
 
-    public ModConfigSpec.BooleanValue corsairKeyboardLighting;
-    
     public ModConfigSpec.BooleanValue cmHeaderTextVisible, cmClassicColors;
-    public ModConfigSpec.IntValue cmTextXOffset, cmXScale, cmXPos, cmSelectedXOffset, cmSubXOffset;
+    public ModConfigSpec.IntValue cmTextXOffset, cmXScale, cmXPos, cmSelectedXOffset, cmSubXOffset, cmEndLWidth, cmEndRWidth, cmHeaderEndLWidth, cmHeaderEndRWidth, cmReactionEndLWidth, cmReactionEndRWidth;
     
     public ModConfigSpec.BooleanValue hpShowHearts;
     public ModConfigSpec.IntValue hpAlarm;
@@ -99,8 +97,38 @@ public class ClientConfig {
 	                .comment("Command Menu Submenu X Offset %")
 	                .translation(KingdomKeys.MODID + ".config.cm_sub_x_offset")
 	                .defineInRange("cmSubXOffset", 100, -1000, 1000);
-	        
-	        builder.pop();
+
+			cmEndLWidth = builder
+					.comment("Command Menu Element Left End Segment Width")
+					.translation(KingdomKeys.MODID + ".config.cm_end_l_width")
+					.defineInRange("cmEndLWidth", 10, 0, 256);
+
+			cmEndRWidth = builder
+				.comment("Command Menu Element Right End Segment Width")
+				.translation(KingdomKeys.MODID + ".config.cm_end_r_width")
+				.defineInRange("cmEndRWidth", 10, 0, 256);
+
+			cmHeaderEndLWidth = builder
+				.comment("Command Menu Header Left End Segment Width")
+				.translation(KingdomKeys.MODID + ".config.cm_header_end_l_width")
+				.defineInRange("cmHeaderEndLWidth", 10, 0, 256);
+
+			cmHeaderEndRWidth = builder
+				.comment("Command Menu Header Right End Segment Width")
+				.translation(KingdomKeys.MODID + ".config.cm_header_end_r_width")
+				.defineInRange("cmHeaderEndRWidth", 15, 0, 256);
+
+			cmReactionEndLWidth = builder
+				.comment("Reaction command left end segment width")
+				.translation(KingdomKeys.MODID + ".config.cm_reaction_end_l_width")
+				.defineInRange("cmReactionEndLWidth", 10, 0, 256);
+
+			cmReactionEndRWidth = builder
+				.comment("Reaction command right end segment width")
+				.translation(KingdomKeys.MODID + ".config.cm_reaction_end_r_width")
+				.defineInRange("cmReactionEndRWidth", 10, 0, 256);
+
+			builder.pop();
 	        
 	        builder.push("hp_bar");